// src/lib/api/models/DrawingStyleResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DrawingStyleResponse = {
    drawingStyleId: number;
    name: string;
    description: string;
    exampleImageUrl?: (string | null);
    disabled: boolean;
};